services/inventory-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
kafka-python==2.0.2
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
python-dotenv==1.0.0
